+++ src/components/alerts/proposalDisplay/ProposalDisplay.test.tsx
@@ -4,11 +4,14 @@ import { StoreProvider } from '@ska-telescope/ska-gui-local-storage';
 import '@testing-library/jest-dom';
 import ProposalDisplay from './ProposalDisplay';
 import { AppFlowProvider } from '@/utils/appFlow/AppFlowContext';
+import { ThemeA11yProvider } from '@/utils/colors/ThemeAllyContext';
 
 const wrapper = (component: React.ReactElement) => {
   return render(
     <StoreProvider>
-      <AppFlowProvider>{component}</AppFlowProvider>
+      <AppFlowProvider>
+        <ThemeA11yProvider>{component}</ThemeA11yProvider>
+      </AppFlowProvider>
     </StoreProvider>
   );
 };
@@ -34,7 +37,7 @@ describe('<ProposalDisplay />', () => {
           cycle: '',
           proposalType: 0,
           proposalSubType: undefined,
-          scienceCategory: null,
+          scienceCategory: 0,
           scienceSubCategory: undefined,
           investigators: undefined,
           abstract: undefined,
@@ -49,7 +52,8 @@ describe('<ProposalDisplay />', () => {
           technicalLoadStatus: undefined,
           dataProductSDP: undefined,
           dataProductSRC: undefined,
-          pipeline: undefined
+          pipeline: undefined,
+          calibrationStrategy: []
         }}
         open={true}
         onClose={mockActionClose}
@@ -89,7 +93,8 @@ describe('<ProposalDisplay />', () => {
           technicalLoadStatus: undefined,
           dataProductSDP: undefined,
           dataProductSRC: undefined,
-          pipeline: undefined
+          pipeline: undefined,
+          calibrationStrategy: []
         }}
         open={true}
         onClose={mockActionClose}