+++ src/components/alerts/validationResults/ValidationResults.test.tsx
@@ -4,11 +4,14 @@ import { StoreProvider } from '@ska-telescope/ska-gui-local-storage';
 import '@testing-library/jest-dom';
 import ValidationResults from './ValidationResults';
 import { AppFlowProvider } from '@/utils/appFlow/AppFlowContext';
+import { ThemeA11yProvider } from '@/utils/colors/ThemeAllyContext';
 
 const wrapper = (component: React.ReactElement) => {
   return render(
     <StoreProvider>
-      <AppFlowProvider>{component}</AppFlowProvider>
+      <AppFlowProvider>
+        <ThemeA11yProvider>{component}</ThemeA11yProvider>
+      </AppFlowProvider>
     </StoreProvider>
   );
 };
@@ -37,7 +40,7 @@ describe('<ValidationResults />', () => {
           cycle: '',
           proposalType: 0,
           proposalSubType: undefined,
-          scienceCategory: null,
+          scienceCategory: 0,
           scienceSubCategory: undefined,
           investigators: undefined,
           abstract: undefined,
@@ -52,7 +55,8 @@ describe('<ValidationResults />', () => {
           technicalLoadStatus: undefined,
           dataProductSDP: undefined,
           dataProductSRC: undefined,
-          pipeline: undefined
+          pipeline: undefined,
+          calibrationStrategy: []
         }}
         results={['LOOKS OK', 'NO ISSUES FOUND']}
       />