+++ CHANGELOG.md
@@ -32,6 +32,7 @@ UNRELEASED
 * Changed: oda to 14.0.4
 * Changed: Adjusted screen layout following removal of help panel 
 * Changed: Update the observation pages so that the addition can be inline if there is a maximum of 1 observation allowed
+* Added: Generate default observation, data product, calibration strategy and sensitivity calculations on target creation
 
 8.0.0
 ----------